installation before zipping is implemented

+++ run_surefire_rule_azure.sh
@@ -66,13 +66,16 @@ fi
 echo "Location of module: $module"
 
 # echo "================Installing the project"
-bash $dir/install-project.sh "$slug" "$MVNOPTIONS" "$USER" "$module" "$sha" "$dir" "$fullTestName" "${RESULTSDIR}"
-ret=${PIPESTATUS[0]}
-mv mvn-install.log ${RESULTSDIR}
-if [[ $ret != 0 ]]; then
-    # mvn install does not compile - return 0
-    echo "Compilation failed. Actual: $ret"
-    exit 1
+projectname=${slug%/*}
+if [[ -! -f "$AZ_BATCH_TASK_WORKING_DIR/input/$projectname.zip" ]]; then
+    bash $dir/install-project.sh "$slug" "$MVNOPTIONS" "$USER" "$module" "$sha" "$dir" "$fullTestName" "${RESULTSDIR}"
+    ret=${PIPESTATUS[0]}
+    mv mvn-install.log ${RESULTSDIR}
+    if [[ $ret != 0 ]]; then
+        # mvn install does not compile - return 0
+        echo "Compilation failed. Actual: $ret"
+        exit 1
+    fi
 fi
 
 cd ~/$slug
